feat: add device utility functions and extcap plumbing for Wireshark integration
